Organize a bit better
